IRLab2/requirements.txt
tqdm==4.45.0
nltk==3.5
numpy==1.18.3
gensim==3.8.2
jieba==0.42.1
pyltp==0.2.1
python_Levenshtein==0.12.0
scikit_learn==0.23.1
